Fragmentation of code for better understanding and linking [C_Error]

+++ KIC-Runtime/hotkey.cpp
@@ -1,10 +1,10 @@
-#include "types.h"
 #include "hotkey.h"
+
 
 template<typename ...args>
 hotkey::hotkey(args... arg) {
-	keys = new uint16_t[std::size(arg)];
-	for (size_t i = 0; i < std::size(arg); i++) {
+	keys = new u16[std::size(arg)];
+	for (u64 i = 0; i < std::size(arg); i++) {
 		keys[i] = std::get<i>(arg);
 	}
 }
